requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-ckeditor==5.4.0
django-contrib-comments==1.8.0
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-fluent-comments==2.0.2
django-js-asset==1.1.0
django-tag-parser==3.1
django-threadedcomments==1.2
djangorestframework==3.8.2
idna==2.6
Markdown==2.6.11
mysqlclient==1.3.12
Pillow==5.1.0
python-akismet==0.4.1
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
